Correctly handle hi-dpi mobiles. (#34)
Media-query are great... Except phones tends to have hi-dpi now,
which leads to the actual resolution of the websites on the phone in
number of pixels to be higher than on desktop. Thus the css media-query
initially made for narrow screen don't trigger.

You now need to put in the HTML an instruction for the browser to "fake"
a css resolution to get the right layout.